Move all items about draw from Team to new class Draw.
