+++ FootballTeamManager/FootballTeamManager.csproj
@@ -302,6 +302,7 @@
     </Compile>
     <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Models\AccountViewModels.cs" />
+    <Compile Include="Models\Draw.cs" />
     <Compile Include="Models\Fixture.cs" />
     <Compile Include="Models\IdentityModels.cs" />
     <Compile Include="Models\ManageViewModels.cs" />
